+++ iPhone/Classes/Login/TestDriveTour_ipad.m
@@ -84,7 +84,7 @@
     }
     //Add flurry
     NSInteger timeInSeconds =  -round( [self.startTime timeIntervalSinceNow]);
-    NSDictionary *dict = @{@"Seconds on Tour":[NSString stringWithFormat:@"%d" ,timeInSeconds]};
+    NSDictionary *dict = @{@"Seconds on Tour":[NSString stringWithFormat:@"%ld" ,(long)timeInSeconds]};
     [Flurry logEvent:@"Test Drive:Tour" withParameters:dict];
     // End timer
     [Flurry endTimedEvent:@"Test Drive:Tour" withParameters:nil];
